+++ src/index.ts
@@ -5,7 +5,7 @@ export {Version} from './version'
 export {err} from './error'
 
 export {gql} from './gql'
-export {GRef as HgRef} from './hgref'
+export {GRef as HgRef} from './gref'
 export {LinkUrl} from './link-url'
 
 import {Schema} from './schema'
