+++ src/linker.ts
@@ -5,7 +5,7 @@ import { Maybe } from 'graphql/jsutils/Maybe'
 import { ImportNode, ImportsParser } from './import'
 import type { IScope } from './scope'
 import {LinkUrl} from './link-url'
-import { GRef } from './hgref'
+import { GRef } from './gref'
 import { scopeNameFor } from './names'
 import { groupBy } from './each'
 import { De, HasGref } from './de'
