+++ src/schema.ts
@@ -4,7 +4,7 @@ import { Maybe } from 'graphql/jsutils/Maybe'
 import { refNodesIn, byRef, Defs, isLocatable, Locatable, fill, De } from './de'
 import { id, Link, Linker, LINK_DIRECTIVES } from './linker'
 import directives from './directives'
-import { GRef } from './hgref'
+import { GRef } from './gref'
 import Scope, { including, IScope } from './scope'
 import { isAst } from './is'
 import gql from './gql'
