+++ src/scope.ts
@@ -2,7 +2,7 @@ import recall, { use } from '@protoplasm/recall'
 import { ASTNode, Kind, visit } from 'graphql'
 import { Linker, type Link } from './linker'
 import { De, Def, Defs, hasRef, isLocatable, isLocated, Locatable, Located } from './de'
-import GRef from './hgref'
+import GRef from './gref'
 import { isAst, hasName } from './is'
 import LinkUrl from './link-url'
 import { getPrefix, scopeNameFor, toPrefixed } from './names'
